requirements.txt
Flask
python-dotenv
google-genai
qdrant-client
pypdf